Revert "Merge branch 'enemy'"
This reverts commit d9f1537fab12f5ce1406c3658a339ffb38c7539a, reversing
changes made to fe96cb8be55a9968fadec3926d1936faa42573e1.

+++ Gun N Hook/game.c
@@ -4,7 +4,6 @@
 #include "mainmenu.h"
 #include "movement.h"
 #include "collision_utils.h"
-#include "game.h"
 
 
 
@@ -17,7 +16,7 @@ void Game_Init(void)
 
 void Game_Update(void)
 {
-	float dt = CP_System_GetDt();
+	double dt = CP_System_GetDt();
 	CP_Graphics_ClearBackground(CP_Color_Create(100, 100, 100, 255));
 
 	if (CP_Input_KeyTriggered(KEY_Q))
