+++ Gun N Hook/movement.c
@@ -1,10 +1,8 @@
 #include "cprocessing.h"
 #include <stdio.h>;
 #include "utils.h";
-#include "game.h"
 
-
-void basic_movement(float* player_x, float* player_y, float* speed_x, float* speed_y, int* on_ground, float dt) {
+void basic_movement(double*player_x, double*player_y, double*speed_x, double*speed_y, int *on_ground, double dt) {
 	if (CP_Input_KeyDown(KEY_D)) {
 		*speed_x = 300;
 		*player_x += *speed_x * dt;
@@ -37,14 +35,14 @@ void basic_movement(float* player_x, float* player_y, float* speed_x, float* spe
 
 }
 
-void gravity(float* player_y, float* speed_y, float dt) {
+void gravity(double* player_y, double *speed_y, double dt) {
 	*speed_y = 500;
 	*player_y += *speed_y * dt;
 }
 
-void drawGrapple(float* player_x, float* player_y, float* grapple_x, float* grapple_y, float dt) {
-	float end_x = 0;
-	float end_y = 0;
+void drawGrapple(double* player_x, double* player_y,double* grapple_x, double* grapple_y, double dt) {
+	double end_x=0;
+	double end_y=0;
 	if (CP_Input_MouseTriggered(MOUSE_BUTTON_RIGHT)) {
 		double mouse_x = CP_Input_GetMouseX();
 		double mouse_y = CP_Input_GetMouseY();
@@ -59,6 +57,6 @@ void drawGrapple(float* player_x, float* player_y, float* grapple_x, float* grap
 		//printf("x is: %lf and y is: %lf\n", end_x, end_y);
 
 	}
-	CP_Settings_Stroke(CP_Color_Create(255, 255, 255, 255));
-	CP_Graphics_DrawLine(*player_x, *player_y, *grapple_x, *grapple_y);
+		CP_Settings_Stroke(CP_Color_Create(255, 255, 255, 255));
+		CP_Graphics_DrawLine(*player_x, *player_y, *grapple_x, *grapple_y);
 }
